Full Simulation of 2 Coupled Harmonic Oscillators added
With this commit, the full pipeline to perform a Suzuki decomposition
of the underlying time evolution operator of the underlying quantum
algorithm for the Hamiltonian simulation of N coupled harmonic
oscillators have been created.

Furthermore, functions required to obtain the analytical solution to
the classical mechanical problem have also been deployed.